protegendo views de usuario nao autenticados
atualizar  vendas e orcamentos nao esta funcionando corretamente

+++ colaborador/views.py
@@ -3,16 +3,21 @@ from django.views.generic import ListView
 
 from colaborador.models import Colaborador
 from colaborador.forms import ColaboradorForms
+from django.contrib.auth.decorators import login_required
 
 
 class ListaColaborador(ListView):
     model = Colaborador
     template_name = 'colaborador/pagina-inicial-colaborador.html'
 
+    def get_queryset(self):
+        return Colaborador.objects.filter(usuario=self.request.user)
 
+
+@login_required
 def cadastrarcolaborador(request):
     template_name = 'colaborador/formularios/formulario-cadastrar-colaborador.html'
-    form = ColaboradorForms(request.POST or None)
+    form = ColaboradorForms(request.POST or None, initial={'usuario': request.user})
 
     if request.method == 'POST':
         if form.is_valid():
@@ -26,10 +31,15 @@ def cadastrarcolaborador(request):
     return render(request, template_name, context)
 
 
+
+@login_required
 def editarcolaborador(request, pk):
     template_name = 'colaborador/formularios/formulario-editar-colaborador.html'
     instance = Colaborador.objects.get(pk=pk)
-    form = ColaboradorForms(request.POST or None, instance=instance)
+    form = ColaboradorForms(request.POST or None, instance=instance, initial={'usuario': request.user})
+
+    if instance.usuario != request.user:
+        raise PermissionError
 
     if request.method == 'POST':
         if form.is_valid():
@@ -43,8 +53,15 @@ def editarcolaborador(request, pk):
     return render(request, template_name, context)
 
 
+@login_required
 def apagarcolaborador(request, pk):
     template_name = 'colaborador/tabela/tabela-colaborador.html'
     obj = Colaborador.objects.get(pk=pk)
-    obj.delete()
+    if obj.usuario == request.user:
+        obj.delete()
+    else:
+        raise PermissionError
     return render(request, template_name)
+
+
+#source /Users/convenienciadigital/Documents/GitHub/PDV-django/venv/bin/activate
